Fix stripe training

+++ tile/codegen/test/load_index.cc
@@ -366,19 +366,19 @@ TEST(LoadIndexTest, SimpleIndex) {
     }
   )***";
   runinfo.input_shapes.emplace("A", SimpleShape(DataType::FLOAT32, {4, 4, 4, 4}));
-  runinfo.output_shapes.emplace("B", SimpleShape(DataType::FLOAT32, {4, 4, 4, 4}));
+  runinfo.output_shapes.emplace("B", SimpleShape(DataType::INT32, {4, 4, 4, 4}));
   auto program = GenerateStripe(runinfo);
   IVLOG(1, "Before stripe optimization: " << *program->entry);
 
   std::string expected_stripe = R"**(0: #program #total_macs=0
     block []:1 ( // load_index_simple
         #user none new@0x00000000 A[0, 0, 0, 0] fp32:I(4, 4, 4, 4):(64, 16, 4, 1):1 KiB
-        #user none new@0x00000000 B[0, 0, 0, 0] fp32:I(4, 4, 4, 4):(64, 16, 4, 1):1 KiB
+        #user none new@0x00000000 B[0, 0, 0, 0] i32:I(4, 4, 4, 4):(64, 16, 4, 1):1 KiB
     ) {
       0: #main 
       block []:1 ( // main
           in A[0, 0, 0, 0] fp32:I(4, 4, 4, 4):(64, 16, 4, 1):1 KiB, E(4, 4, 4, 4):1 KiB
-          out B[0, 0, 0, 0]:assign fp32:I(4, 4, 4, 4):(64, 16, 4, 1):1 KiB, E(4, 4, 4, 4):1 KiB
+          out B[0, 0, 0, 0]:assign i32:I(4, 4, 4, 4):(64, 16, 4, 1):1 KiB, E(4, 4, 4, 4):1 KiB
       ) {
         0: #eltwise #eltwise_index #kernel 
         block [i1:4, i2:4, i3:4, i4:4]:256 ( // kernel_0(A)
@@ -427,19 +427,19 @@ TEST(LoadIndexTest, AffineIndex) {
     }
   )***";
   runinfo.input_shapes.emplace("A", SimpleShape(DataType::FLOAT32, {8, 8, 256, 8}));
-  runinfo.output_shapes.emplace("B", SimpleShape(DataType::FLOAT32, {8, 8, 256, 8}));
+  runinfo.output_shapes.emplace("B", SimpleShape(DataType::INT32, {8, 8, 256, 8}));
   auto program = GenerateStripe(runinfo);
   IVLOG(1, "Before stripe optimization: " << *program->entry);
 
   std::string expected_stripe = R"**(0: #program #total_macs=0
     block []:1 ( // load_index_affine
         #user none new@0x00000000 A[0, 0, 0, 0] fp32:I(8, 8, 256, 8):(16384, 2048, 8, 1):512 KiB
-        #user none new@0x00000000 B[0, 0, 0, 0] fp32:I(8, 8, 256, 8):(16384, 2048, 8, 1):512 KiB
+        #user none new@0x00000000 B[0, 0, 0, 0] i32:I(8, 8, 256, 8):(16384, 2048, 8, 1):512 KiB
     ) {
       0: #main 
       block []:1 ( // main
           in A[0, 0, 0, 0] fp32:I(8, 8, 256, 8):(16384, 2048, 8, 1):512 KiB, E(8, 8, 256, 8):512 KiB
-          out B[0, 0, 0, 0]:assign fp32:I(8, 8, 256, 8):(16384, 2048, 8, 1):512 KiB, E(8, 8, 256, 8):512 KiB
+          out B[0, 0, 0, 0]:assign i32:I(8, 8, 256, 8):(16384, 2048, 8, 1):512 KiB, E(8, 8, 256, 8):512 KiB
       ) {
         0: #eltwise #eltwise_index #kernel 
         block [i1:8, i2:8, i3:256, i4:8]:131072 ( // kernel_0(A)
@@ -490,25 +490,25 @@ TEST(LoadIndexTest, MultiLoadIndex) {
     }
   )***";
   runinfo.input_shapes.emplace("A", SimpleShape(DataType::FLOAT32, {4, 4, 256, 8}));
-  runinfo.output_shapes.emplace("B", SimpleShape(DataType::FLOAT32, {4, 4, 256, 8}));
-  runinfo.output_shapes.emplace("C", SimpleShape(DataType::FLOAT32, {4, 4, 256, 8}));
-  runinfo.output_shapes.emplace("D", SimpleShape(DataType::FLOAT32, {4, 4, 256, 8}));
+  runinfo.output_shapes.emplace("B", SimpleShape(DataType::INT32, {4, 4, 256, 8}));
+  runinfo.output_shapes.emplace("C", SimpleShape(DataType::INT32, {4, 4, 256, 8}));
+  runinfo.output_shapes.emplace("D", SimpleShape(DataType::INT32, {4, 4, 256, 8}));
   auto program = GenerateStripe(runinfo);
   IVLOG(1, "Before stripe optimization: " << *program->entry);
 
   std::string expected_stripe = R"**(0: #program #total_macs=0
     block []:1 ( // load_index_affine
         #user none new@0x00000000 A[0, 0, 0, 0] fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB
-        #user none new@0x00000000 B[0, 0, 0, 0] fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB
-        #user none new@0x00000000 C[0, 0, 0, 0] fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB
-        #user none new@0x00000000 D[0, 0, 0, 0] fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB
+        #user none new@0x00000000 B[0, 0, 0, 0] i32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB
+        #user none new@0x00000000 C[0, 0, 0, 0] i32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB
+        #user none new@0x00000000 D[0, 0, 0, 0] i32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB
     ) {
       0: #main 
       block []:1 ( // main
           in A[0, 0, 0, 0] fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB, E(4, 4, 256, 8):128 KiB
-          out B[0, 0, 0, 0]:assign fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB, E(4, 4, 256, 8):128 KiB
-          out C[0, 0, 0, 0]:assign fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB, E(4, 4, 256, 8):128 KiB
-          out D[0, 0, 0, 0]:assign fp32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB, E(4, 4, 256, 8):128 KiB
+          out B[0, 0, 0, 0]:assign i32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB, E(4, 4, 256, 8):128 KiB
+          out C[0, 0, 0, 0]:assign i32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB, E(4, 4, 256, 8):128 KiB
+          out D[0, 0, 0, 0]:assign i32:I(4, 4, 256, 8):(8192, 2048, 8, 1):128 KiB, E(4, 4, 256, 8):128 KiB
       ) {
         0: #eltwise #eltwise_index #kernel 
         block [i1:4, i2:4, i3:256, i4:8]:32768 ( // kernel_0(A)
@@ -578,7 +578,7 @@ TEST(LoadIndexTest, FuseIndex) {
     }
   )***";
   runinfo.input_shapes.emplace("A", SimpleShape(DataType::FLOAT32, {4, 4, 4, 8}));
-  runinfo.output_shapes.emplace("B", SimpleShape(DataType::FLOAT32, {4, 4, 4, 8}));
+  runinfo.output_shapes.emplace("B", SimpleShape(DataType::INT32, {4, 4, 4, 8}));
   runinfo.output_shapes.emplace("C", SimpleShape(DataType::FLOAT32, {4, 4, 4, 8}));
   auto program = GenerateStripe(runinfo);
   IVLOG(1, "Before stripe optimization: " << *program->entry);
@@ -586,13 +586,13 @@ TEST(LoadIndexTest, FuseIndex) {
   std::string expected_stripe = R"**(0: #program #total_macs=0
     block []:1 ( // load_index_affine
         #user none new@0x00000000 A[0, 0, 0, 0] fp32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB
-        #user none new@0x00000000 B[0, 0, 0, 0] fp32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB
+        #user none new@0x00000000 B[0, 0, 0, 0] i32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB
         #user none new@0x00000000 C[0, 0, 0, 0] fp32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB
     ) {
       0: #main 
       block []:1 ( // main
           in A[0, 0, 0, 0] fp32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB, E(4, 4, 4, 8):2 KiB
-          out B[0, 0, 0, 0]:assign fp32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB, E(4, 4, 4, 8):2 KiB
+          out B[0, 0, 0, 0]:assign i32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB, E(4, 4, 4, 8):2 KiB
           out C[0, 0, 0, 0]:assign fp32:I(4, 4, 4, 8):(128, 32, 8, 1):2 KiB, E(4, 4, 4, 8):2 KiB
       ) {
         0: #eltwise #eltwise_index #kernel
